Disband components in separate files
